refactor: InputFile.web.js
